add note about npm version to readme
there are a number of issues that report being unable to install with newer npm versions. the config in package.json only produces a warning when running with a newer npm version, which is drowned out by deprecation notice for other packages. adding it to the readme at least puts a reference where people would typically start.